move config saving into its own file
